Added option to show popup on a specific display
This is useful, because when the process runs as a post-commit hook or
on a cron job, the user that runs the command might be not the same as
the current user logged in.
This allows to set the display as: DISPLAY=:0.0 (or any other display
you want), so that the notify-send pop-up shows on the currently
logged in user display.

the user can now define their notifier in their config as:
// create a DBus notifier (for Linux)
$notifier = new Sismo\DBusNotifier('DISPLAY=:0.0');